Fixed return value of field getters in case a field is not existing and moved VR functions in the supevisor.py file
